Make default log file mode 0600
When a file is not pre-created, the default mode were 0644. This
could be very problematic for audit log files which require tighter
permissions.

This changes the default mode to be 0600 which is more
restrictive. Note that rotation of log files already uses this
mode by default. Also note that when files are pre-created, the mode the
file already had is respected.